Refactored SocketClient, and removed GraphQLSocket class.
